backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Amazon KDP Navigator — backend API"
readme = "README.md"
authors = [
    { email = "makoto.sugano@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "pyotp>=2.9.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn[standard]>=0.42.0",
]

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"
